pyproject.toml
[project]
name = "chatbot"
version = "0.1.0"
description = "End-to-End Q&A Chatbot using LangChain and OpenAI"
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
    "langchain>=0.3.26",
    "langchain-openai>=0.1.6",
    "llama-index>=0.10.28",
    "python-dotenv",
    "streamlit>=1.26.0",
    "langsmith",
    "langchain-community>=0.0.30",
    "langchain-core>=0.1.48",
    "langserve",
    "langchain-groq>=0.3.6",
    "google-search-results>=2.4.2",
]
